fix(BfsOcclusion): 使用long计算避免整数溢出
feat(VideoOptionsRegistry): 实现选项变更管理和监听功能

实现视频选项注册表的完整功能，包括：
- 选项修改标记和批量应用
- 变更监听器机制
- 选项页面注册
- 重置选项到默认值
- 未保存修改检查

+++ common/src/main/java/com/renderium/accel/BfsOcclusion.java
@@ -121,8 +121,9 @@ public final class BfsOcclusion {
             // 保守估计: Minecraft 渲染距离 32 区块 ≈ 32768 区块截面
             // 使用 65536 作为安全上限，对应 resultSize = 16 + 65536/8 = 8200 bytes
             int estimatedNodeCount = Math.max(8192, (int)(renderDistance * renderDistance * 4));
-            int bitmapSizeWords = (estimatedNodeCount + 31) / 32;
-            int resultSize = 16 + bitmapSizeWords * 4;  // VisibilityResult header(16) + bitmap
+            // 使用 long 计算避免整数溢出，然后安全转换为 int
+            int bitmapSizeWords = Math.toIntExact((estimatedNodeCount + 31L) / 32);
+            int resultSize = 16 + Math.toIntExact((long) bitmapSizeWords * 4);
 
             MemorySegment result = arena.allocate(resultSize);
